fix float pow bug

+++ src/plssvm/backends/CUDA/q_kernel.cu
@@ -27,7 +27,7 @@ __global__ void device_kernel_q_poly(real_type *q, const real_type *data_d, cons
     for (int i = 0; i < num_cols; ++i) {
         temp += data_d[i * num_rows + index] * data_last[i];
     }
-    q[index] = pow(gamma * temp + coef0, degree);
+    q[index] = pow(gamma * temp + coef0, static_cast<int>(degree));
 }
 template __global__ void device_kernel_q_poly(float *, const float *, const float *, const int, const int, const float, const float, const float);
 template __global__ void device_kernel_q_poly(double *, const double *, const double *, const int, const int, const double, const double, const double);
